Edited Model
